feat(debug): Add Fastboot debug logging, Windows USB refactor
Introduce FastbootDebug and sprinkle debug logs across FastbootUtil (RawCommand, GetVarAll) and LibUsbDevice to aid troubleshooting. Simplify FastbootCLI argument parsing, error handling, and device selection; make FastbootUtil IDisposable and ensure proper transport disposal. Refactor Windows USB stack: add CTL_CODE helper, fix P/Invoke signatures (uint sizes), rework LegacyUsbDevice and WinUSBDevice handles/IO, and rewrite WinUSBFinder to robustly enumerate and probe device interfaces. Also update project metadata in the csproj.

+++ FirmwareKit.Comm.Fastboot/Usb/Windows/WinUSBAPI.cs
@@ -47,11 +47,11 @@ public class WinUSBAPI
 
     [DllImport("Winusb.dll", CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall, SetLastError = true)]
     public static extern bool WinUsb_WritePipe(IntPtr DeviceHandle, byte pipeID, byte[] buffer,
-        ulong bufferLen, out ulong bytesTransfered, IntPtr overlapp);
+        uint bufferLen, out uint bytesTransfered, IntPtr overlapp);
 
     [DllImport("Winusb.dll", CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall, SetLastError = true)]
     public static extern bool WinUsb_ReadPipe(IntPtr DeviceHandle, byte pipeID, byte[] buffer,
-        ulong bufferLen, out ulong bytesTransfered, IntPtr overlapp);
+        uint bufferLen, out uint bytesTransfered, IntPtr overlapp);
 
     [DllImport("Winusb.dll", CharSet = CharSet.Auto, CallingConvention = CallingConvention.StdCall, SetLastError = true)]
     public static extern bool WinUsb_ResetPipe(IntPtr InterfaceHandle, byte PipeID);
